==> docker/COBtree/R_scripts/STEP2_single_cell_sampling.R <==
source("/COBtree/R_functions/utils.R")

############## Read input parameters
suppressPackageStartupMessages(library("optparse"))

option_list <- list(
  make_option(c("-t", "--topology"), type="character", default=NULL, help="RDS file path with the topology generated by STEP1 or STEP2",
              metavar="string"),
  make_option(c("-n", "--single_cells"), type="integer", default=1000, 
              help="Number of single cells sampled from the possible genotypes defined from the topology. [default %default]",
              metavar="number"),
  make_option(c("-r", "--shuffle_cols"), type="logical", default = TRUE,
              help="Dataset generated will have a random order of the colums (mutations) [default %default]",
              metavar="string"),
  make_option(c("-p", "--increase_leaf_prob"), type="double", default = 5,
              help="Multiply the probability of sampling from terminal clones (leaf) by the specified values. [default x %default]",
              metavar="string"),
  make_option(c("-o", "--output_file"), type="character", default = "",
              help="Path of the output file. It will include the topology and the single cell genotypes (Dgt)",
              metavar="string"),
  make_option(c("-s", "--seed"), type="integer", default = sample(1:10000, 1),
              help="Seed for random number, is used for reproducibility. [default: random]",
              metavar="number")
)

opt <- parse_args(OptionParser(option_list=option_list))

if(file.exists(opt$topology)) {
  GT <- readRDS(opt$topology)
} else {
  stop("Please, indicate a correct RDS file")
}

GT <- readRDS(opt$topology)

sc_num <- opt$single_cells

clonal_genotypes <- GT$clonal_genotypes

if(opt$output_file == "") {
  out_fn <- gsub(x = opt$topology, 
                 pattern = '.rds', 
                 replacement = paste0('_sc_',sc_num,'.rds'))
} else {
  out_fn <- opt$output_file
  if(!dir.exists(dirname(out_fn))) {
    dir.create(dirname(out_fn), recursive = T)
  }
}

set.seed(opt$seed)

if(opt$increase_leaf_prob != 1) {
  
  adj_m <- as.adj.matrix(B = GT$Bgt)
  leaf_nodes <- rownames(adj_m)[which(rowSums(adj_m)==0)]
  clone_idxs <- c(which(rownames(clonal_genotypes) %in% leaf_nodes),
                       which(!rownames(clonal_genotypes) %in% leaf_nodes))
  
  w_non_leaf = length(clone_idxs) - length(leaf_nodes)
  w_leaf = opt$increase_leaf_prob*w_non_leaf
  
  clone_prob = c(rep(x = (w_leaf/length(leaf_nodes)), length = length(leaf_nodes)),
                 rep(x = 1, length = (length(clone_idxs) - length(leaf_nodes))))
  
  idx_clone = sample(x = clone_idxs, size = sc_num, replace = T, prob = clone_prob/sum(clone_prob))
  
} else {
  idx_clone = sample(x = 1:nrow(clonal_genotypes), size = sc_num, replace = T)
}

idx_col <- 1:ncol(clonal_genotypes)

if(opt$shuffle_cols) {
  idx_col <- sample(x = idx_col, size = length(idx_col), replace = F)
} 

Dgt <- (clonal_genotypes[idx_clone,idx_col])

GT$Dgt <- Dgt
GT$CloneID <- idx_clone

saveRDS(object = GT, file = out_fn)
