==> docker/COBtree/R_scripts/STEP3_make_input_dataset.R <==
############## Read input parameters
suppressPackageStartupMessages(library("optparse"))

option_list <- list(
  make_option(c("-c", "--sampled_cells"), type="character", default=NULL, 
              help="RDS file path with the sampled single cell generated by STEP3",
              metavar="string"),
  
  make_option(c("-p", "--false_positive"), type="double", default=0.005, 
              help="Set the false positive rate. [default %default]",
              metavar="number"),
  
  make_option(c("-n", "--false_negative"), type="double", default=0.05, 
              help="Set the false negative rate. [default %default]",
              metavar="number"),
  
  make_option(c("-m", "--missing_rate"), type="double", default=0.1, 
              help="Set the missing value rate. [default %default]",
              metavar="number"),
  
  make_option(c("-v", "--missing_value"), type="integer", default=3, 
              help="Set the value to represent a missing entry. [default %default]",
              metavar="number"),
  
  make_option(c("-t", "--cell_col"), type="logical", default=TRUE, 
              help="If TRUE the resulting dataset will have mutations (rows) X cells (columns); if FALSE it will be transposed.  [default %default]",
              metavar="number"),
  
  make_option(c("-q", "--sep"), type="character", default = '\t',
              help="Specify the separator. [default %default]",
              metavar="string"),
  
  make_option(c("-o", "--output_file"), type="character", default = "",
              help="Path of the csv file.",
              metavar="string"),
  
  make_option(c("-s", "--seed"), type="integer", default = sample(1:10000, 1),
              help="Seed for random number, is used for reproducibility. [default: random]",
              metavar="number")
)

opt <- parse_args(OptionParser(option_list=option_list))

if(file.exists(opt$sampled_cells)) {
  GT <- readRDS(opt$sampled_cells)
} else {
  stop("Please, indicate a correct RDS file")
}

GT <- readRDS(opt$sampled_cells)

clonal_genotypes <- GT$clonal_genotypes

alpha <- opt$false_positive
beta <- opt$false_negative
missing <- opt$missing_rate 

if(opt$output_file == "") {
  out_fn <- gsub(x = opt$sampled_cells, 
                 pattern = '.rds', 
                 replacement = paste0('_FP_',format(x = alpha, scientific = T),
                                      '_FN_',format(x = beta, scientific = T),
                                      '_M_',format(x = missing, scientific = T),
                                      '.csv'))
} else {
  out_fn <- opt$output_file
  if(!dir.exists(dirname(out_fn))) {
    dir.create(dirname(out_fn), recursive = T)
  }
}

print(out_fn)

set.seed(opt$seed)

Dgt <- GT$Dgt

mv <- opt$missing_value

# add noise
D <- Dgt
idx_1 <- which(D==1)
idx_0 <- which(D==0)

idx_1_0 <- sample(x = idx_1, size = round(length(idx_1)*beta))
idx_0_1 <- sample(x = idx_0, size = round(length(idx_0)*alpha))

D[idx_1_0] <- 0
D[idx_0_1] <- 1

# add missing data
idx_NA <- sample(x = prod(dim(D)), size = round(prod(dim(D))*missing))
D[idx_NA] <- mv

#if(opt$mutation_names_file != "") {
#  cat(colnames(D), 
#      file = opt$mutation_names_file, 
#      sep = '\n')
#}

if(opt$cell_col) {
  D <- t(D)
}

write.table(x = D, 
            file = out_fn, 
            sep = opt$sep, 
            quote = FALSE, 
            row.names = FALSE,
            col.names = FALSE
            )
