==> docker/COBtree/R_scripts/STEP4_compute_SCITE_cons_tree.R <==
####################### Function to compute the logLikelihood using SCITE method #####################################
getLogLikelihood_scite <- function(adjM = NULL, GTcons = NULL, dataset_fn = NULL, nCells = 1000, FP = NULL, FN = NULL){
  
  adjM[which(adjM > 0)] <- 1
  
  nMut <- ncol(adjM)-1
  
  parentVector <- vector(mode = 'numeric', length = nMut)
  for(i in 1:nMut) {
    idx_child_i <- which(colnames(adjM) == i)
    par_i <- rownames(adjM)[which(adjM[,idx_child_i]==1)]
    if(par_i == 'r') {
      parentVector[i] <- nMut+1
    } else {
      parentVector[i] <- which(colnames(GTcons$Dgt) == par_i)
    }
  }
  names(parentVector)<-as.character(1:nMut)
  parentVector_ord <- parentVector[as.numeric(colnames(GTcons$Dgt))]
  PVTree <- paste0(parentVector_ord, collapse = ',')
  
  strLogLikelihood <- system(paste0("/COBtree/SCITElik/getSCITEloglik -i \"", dataset_fn, "\"", 
                                    " -n ", nMut, 
                                    " -m ", nCells,
                                    " -fd ", FP,
                                    " -ad ", FN,
                                    " -t ", PVTree), 
                             intern = TRUE)
  
  logLik_r <- as.numeric(strsplit(x = strLogLikelihood, split = ':')[[1]][2])
  results <- list(logLik = logLik_r, PVTree = parentVector_ord)
  return(results)
  
}


############## Read input parameters
suppressPackageStartupMessages(library("optparse"))

option_list <- list(
  make_option(c("-g", "--ground_truth"), type="character", default=NULL, 
              help="RDS file path with the ground truth data generated by STEP3",
              metavar="string"),
              
  make_option(c("-d", "--dataset"), type="character", default=NULL, 
              help="CSV file path with the single cell dataset generated by STEP3",
              metavar="string"),
                
  make_option(c("-s", "--scite_samples"), type="character", default=NULL, 
              help=".samples file path generated by SCITE. It includes the sampled trees",
              metavar="string"),
  
  make_option(c("-l", "--scite_ml0"), type="character", default=NULL, 
              help="ml0.gv file path generated by SCITE. It includes the maximum likelihood tree",
              metavar="string"),
  
  make_option(c("-p", "--likelihood_thr"), type="character", default=0.2, 
              help="Specify the min likelihood threshold to filter out sampled tree (represented as a percentual increase of the maximum likelihood obtained). [default %default]",
              metavar="string"),
  
  make_option(c("-o", "--output_file"), type="character", default = "",
              help="Path of the rds file which includes ground truth data plus consensus tree computed.",
              metavar="string")
)

opt <- parse_args(OptionParser(option_list=option_list))

if(file.exists(opt$ground_truth)) {
  GT <- readRDS(opt$ground_truth)
} else {
  stop("Please, indicate a correct RDS file")
}

if(file.exists(opt$dataset)) {
  dataset_fn <- opt$dataset
}else{
  stop("Please, indicate a correct CSV file")
}

if(file.exists(opt$scite_samples)) {
  scite_samples_fn <- opt$scite_samples
} else {
  stop("Please, indicate a correct .samples file")
}

if(file.exists(opt$scite_ml0)) {
  scite_ml_fn <- opt$scite_ml0
} else {
  stop("Please, indicate a correct _ml0.gv file")
}


if(opt$output_file == "") {
  out_fn <- gsub(x = opt$scite_samples, 
                 pattern = '.samples$', 
                 replacement = '_consensus.rds')
} else {
  out_fn <- opt$output_rds_file
  if(!dir.exists(dirname(out_fn))) {
    dir.create(dirname(out_fn), recursive = T)
  }
}

source("/COBtree/R_functions/consensusTreeFunctions.R")
source("/COBtree/R_functions/utils.R")

suppressMessages(library(graph))

nMut <- ncol(GT$clonal_genotypes)

run_info <- strsplit(x=gsub(pattern = ".samples$", replacement = "", x = scite_samples_fn), 
                     split="_",fixed=T)[[1]]

branch = GT$branch

alpha <- as.numeric(run_info[which(run_info == "FP") + 1])
beta <- as.numeric(run_info[which(run_info == "FN") + 1])
miss <- as.numeric(run_info[which(run_info == "M") + 1])

nCells <- as.numeric(run_info[which(run_info == "sc") + 1])

MCMC <- as.numeric(run_info[which(run_info == "MCMC") + 1])


lik_thr_V = as.numeric(unlist(strsplit(x = opt$likelihood_thr, split = ',', fixed = T)))

adj_scite_ml_i <- scite.2.adjM(scite_ml0_fn = scite_ml_fn, mutNames = colnames(GT$Dgt))
GT$SCITE_ML_adj <- adj_scite_ml_i

adj_Bgt <- as.adj.matrix(GT$Bgt, sorting=T)


GT$consensus_tree <- list()

GT$metrics <- data.frame(SCITEfn = character(), 
                    n_mut = numeric(),
                    branch = numeric(),
                    MCMCsteps = numeric(),
                    alpha = numeric(),
                    beta = numeric(),
                    missing = numeric(),
                    GT_logLik = numeric(),
                    ML_logLik = numeric(),
                    COB_logLik = numeric(),
                    ML_CGerrors = numeric(),
                    COB_CGerrors = numeric(),
                    ML_PCdist = numeric(),
                    COB_PCdist = numeric(),
                    sampled_trees = numeric(),
                    lik_thr = numeric())
                    
GT_logLik = getLogLikelihood_scite(adjM = as.adj.matrix(B = GT$Bgt, sorting = T),
                                    GTcons = GT, 
                                    dataset_fn = dataset_fn, 
                                    nCells = nCells, 
                                    FP = alpha, 
                                    FN = beta)
                                    
ML_logLik = getLogLikelihood_scite(adjM = GT$SCITE_ML_adj,
                                    GTcons = GT, 
                                    dataset_fn = dataset_fn, 
                                    nCells = nCells, 
                                    FP = alpha, 
                                    FN = beta)


############ Compute Clonal Genotype errors for scite ML tree
GT_clonal_geno <- getClonalGenotypes_sorted(B = GT$Bgt)
ML_clonal_geno <- getClonalGenotypes_sorted(B = as.B.rooted(GT$SCITE_ML_adj))
ML_CGerrors = sum(abs(GT_clonal_geno - ML_clonal_geno))

############ Compute Parent-Child distance for scite ML tree
ML_PCdist = distance.parent_child.SCITEsamples(pvA = GT_logLik$PVTree, pvB = ML_logLik$PVTree)


for(lthr in lik_thr_V){

consensus_tree_i <-  computeMST.fromSCITEsamples(scite_samples_fn=scite_samples_fn,
                                                 logLik_thr = lthr,
                                                 mut_names=colnames(GT$Dgt), 
                                                 tree_every = 1,
                                                 tree_buffer=Inf,
                                                 verbose=F)

consensus_tree_i$lik_thr = lthr

COB_logLik_i = getLogLikelihood_scite(adjM = consensus_tree_i$cob_weigth_adj,
                                    GTcons = GT, 
                                    dataset_fn = dataset_fn, 
                                    nCells = nCells, 
                                    FP = alpha, 
                                    FN = beta)

COB_clonal_geno <- getClonalGenotypes_sorted(B = consensus_tree_i$cob_tree)
COB_CGerrors_i <- sum(abs(GT_clonal_geno - COB_clonal_geno))
COB_PC_dist_i <- distance.parent_child.SCITEsamples(pvA = GT_logLik$PVTree, pvB = COB_logLik_i$PVTree)

GT$consensus_tree[[paste0('lik_thr', lthr)]] <- consensus_tree_i 

adj_COBtree_i <- as.adj.matrix(consensus_tree_i$cob_tree, sorting=T)

GT$metrics <- rbind(GT$metrics, data.frame(SCITEfn = scite_samples_fn, 
                                            n_mut = nMut,
                                            branch = branch,
                                            MCMCsteps = MCMC,
                                            alpha = alpha,
                                            beta = beta,
                                            missing = miss,
                                            GT_logLik = GT_logLik$logLik,
                                            ML_logLik = ML_logLik$logLik,
                                            COB_logLik = COB_logLik_i$logLik,
                                            ML_CGerrors = ML_CGerrors,
                                            COB_CGerrors = COB_CGerrors_i,
                                            ML_PCdist = ML_PCdist,
                                            COB_PCdist = COB_PC_dist_i,
                                            sampled_trees = consensus_tree_i$sampled_trees,
                                            lik_thr = lthr))

}

saveRDS(object = GT, file = out_fn)
write.table(x = GT$metrics, file = gsub(x=out_fn,pattern='_consensus.rds',replacement='_metrics.tsv'), sep='\t',row.names = F, col.names=T, quote=F)
